Machine learning component, 2 of 3 pieces of code.
This code picks up the output.CSV from piece one of the code. 
This CSV is read in as infile. It is given headers and the features named. 
The independent variables (features) are named for the model. (X)
The dependent variable (target) is also identified for the model. (y)
The data is split for training and testing (70/30%). 
Linear regression model is used. 
Second main piece is the 'importance' coefficient, this is a value between -1 and 1. This identifies which feature has the biggest effect on the prediction. 
This is built into a dictionary for the third piece as well as plotted on a graph. 
The dictionary output is networkattributes